% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_prompts.r
\name{annotate_code}
\alias{annotate_code}
\title{Annotate R code with inline comments}
\usage{
annotate_code(code, ...)
}
\arguments{
\item{code}{A character vector of R code. If missing the code currently
selected in RStudio is documented (If RStudio is used).}

\item{...}{passed on to \code{\link{askgpt}}.}
}
\value{
A character vector.
}
\description{
Annotate R code with inline comments
}
