% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askgpt.r
\name{askgpt}
\alias{askgpt}
\title{Ask openai's GPT models a question}
\usage{
askgpt(prompt, return_answer = FALSE)
}
\arguments{
\item{prompt}{What you want to ask}

\item{return_answer}{Should the answer be returned as an object instead of
printing it to the screen?}
}
\description{
Ask openai's GPT models a question
}
\examples{
\dontrun{
askgpt("What is an R function?")
askgpt("What is wrong with my last command?")
askgpt("Can you help me with the function aes() from ggplot2?")
}
}
