% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{completions_api}
\alias{completions_api}
\title{Request answer from openai's completions API}
\usage{
completions_api(
  prompt,
  model = NULL,
  temperature = 0.2,
  max_tokens = 2048L,
  stream = FALSE,
  api_key = NULL,
  ...
)
}
\arguments{
\item{prompt}{character string of the prompt to be completed.}

\item{model}{character string of the model to be used (defaults to
"text-davinci-003").}

\item{temperature}{numeric value between 0 and 1 to control the randomness of
the output (defaults to 0.2; lower values like 0.2 will make answers more
focused and deterministic).}

\item{max_tokens}{The maximum number of tokens to generate in the completion.
2048L is the maximum the models accept.}

\item{stream}{logical value to indicate whether the response should be
streamed.}

\item{api_key}{set the API key. If NULL, looks for the env OPENAI_API_KEY.}

\item{...}{additional parameters to be passed to the API (see [the API
documentation](https://platform.openai.com/docs/api-reference/completions)}
}
\description{
Request answer from openai's completions API
}
\examples{
\dontrun{
completions_api("The quick brown fox")
}
}
