% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{estimate_token}
\alias{estimate_token}
\title{Estimate token count}
\usage{
estimate_token(x, mult = 1.6)
}
\arguments{
\item{x}{character vector}

\item{mult}{the multiplier used}
}
\value{
a integer vector of token counts
}
\description{
Estimate token count
}
\details{
This function estimates how many tokens the API will make of the
input words. For the models 1 word is more than one token. The default
multiplier value resulted from testing the API. See
<https://help.openai.com/en/articles/4936856-what-are-tokens-and-how-to-count-them>
for more information.
}
\examples{
estimate_token("this is a test")
}
