% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_addins.r
\name{improve_addin}
\alias{improve_addin}
\title{Improve code/documentation/writing using a prompt}
\usage{
improve_addin()
}
\value{
No return value, opens a new file in RStudio
}
\description{
`tutorialise_addin()` opens an [RStudio
  gadget](https://shiny.rstudio.com/articles/gadgets.html) and
  [addin](http://rstudio.github.io/rstudioaddins/) that can be used to
  improve existing code, documentation, or writing.
}
