% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.r
\name{list_models}
\alias{list_models}
\title{List Models}
\usage{
list_models(api_key = NULL)
}
\arguments{
\item{api_key}{set the API key. If NULL, looks for the env OPENAI_API_KEY.}
}
\value{
A tibble with available models
}
\description{
List the models available in the API. You can refer to the [Models
documentation](https://platform.openai.com/docs/models) to understand what
models are available and the differences between them.
}
\examples{
\dontrun{
completions_api("The quick brown fox")
}
}
