% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{log_init}
\alias{log_init}
\title{Initiate error logging}
\usage{
log_init(...)
}
\arguments{
\item{...}{forwarded to \code{\link[rlang]{global_entrace}}}
}
\value{
No return value, called to enable rlang error logging
}
\description{
Initiate error logging
}
\details{
Just an alias for rlang::global_entrace() with a more fitting name (for the
purpose here).
}
