% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{new_conversation}
\alias{new_conversation}
\title{Start a new conversation}
\usage{
new_conversation()
}
\value{
Does not return a value
}
\description{
Deletes the local prompt and response history to start a new conversation.
}
