% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{parse_response}
\alias{parse_response}
\title{Parse response from API functions}
\usage{
parse_response(response)
}
\arguments{
\item{response}{a response object from \code{\link{chat_api}} or
\code{\link{completions_api}}}
}
\value{
a character vector
}
\description{
Parse response from API functions
}
