% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{prompt_history}
\alias{prompt_history}
\title{Return the prompt/response history}
\usage{
prompt_history(n = Inf)
}
\arguments{
\item{n}{number of prompts/responses to return.}
}
\value{
a character vector
}
\description{
Return the prompt/response history
}
