% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_prompts.r
\name{test_function}
\alias{test_function}
\title{Test R code}
\usage{
test_function(code, ...)
}
\arguments{
\item{code}{A character vector of R code. If missing the code currently
selected in RStudio is documented (If RStudio is used).}

\item{...}{passed on to \code{\link{askgpt}}.}
}
\value{
A character vector.
}
\description{
Test R code
}
