% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.r
\docType{data}
\name{token_limits}
\alias{token_limits}
\title{Max tokens limits of the different models}
\format{
An object of class \code{data.frame} with 6 rows and 2 columns.
}
\source{
<https://platform.openai.com/docs/models/overview>
}
\usage{
token_limits
}
\description{
OpenAI's token limits for different models.
}
\keyword{datasets}
