% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tutorial_addin.r
\name{tutorialise_addin}
\alias{tutorialise_addin}
\title{Turn R code into a tutorial}
\usage{
tutorialise_addin()
}
\value{
No return value, opens a new file in RStudio
}
\description{
`tutorialise_addin()` opens an [RStudio
  gadget](https://shiny.rstudio.com/articles/gadgets.html) and
  [addin](http://rstudio.github.io/rstudioaddins/) that turns selected code
  into an R Markdown/Quarto Tutorial.
}
