% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askgpt.r
\name{askgpt}
\alias{askgpt}
\title{Ask openai's GPT models a question}
\usage{
askgpt(
  prompt,
  chat = TRUE,
  stream = FALSE,
  progress = TRUE,
  return_answer = FALSE,
  ...
)
}
\arguments{
\item{prompt}{What you want to ask}

\item{chat}{whether to use the chat API (i.e., the same model as ChatGPT) or
the completions API.}

\item{stream}{Return pieces of the answer to the screen instead of waiting
for the request to be completed.}

\item{progress}{Show a progress spinner while the request to the API has not
been fulfilled.}

\item{return_answer}{Should the answer be returned as an object instead of
printing it to the screen?}

\item{...}{additional options forwarded to \code{\link{chat_api}} or
\code{\link{completions_api}} respectively.}
}
\value{
either an httr2 response from one of the APIs or a character vector
  (if return_answer).
}
\description{
Ask openai's GPT models a question
}
\examples{
\dontrun{
askgpt("What is an R function?")
askgpt("What is wrong with my last command?")
askgpt("Can you help me with the function aes() from ggplot2?")
}
}
